Consider checked exceptions.
The ConcurrentPool needs to perform some cleanups when an exception
occurs in the user-provided implementation of PoolObjectFactory.

PoolObjectFactory methods do not declare any thrown Exceptions,
so it is sufficient to catch Error and RuntimeException in the
ConcurrentPool.

This stops working when one uses this library from kotlin:
It is possible to throw a checked exception from a kotlin implementation
of PoolObjectFactory, and when that happens the cleanups don't occur in
the ConcurrentPool, slowly exhausting slots in the pool and leading to
object starvation.

This commit modifies the catch clauses in the ConcurrentPool and
SamplingPoolReducer to handle any kind of Throwable.